refactor(docs): collapse 7 section route dirs into one [section]/[sub]
The seven section folders each had near-identical layout.tsx / page.tsx /
[sub]/page.tsx (21 files). Collapse to a single dynamic route:

- app/[section]/layout.tsx — async, awaits params, getSection→notFound
  (replaces the seven getSection('x')! non-null assertions, a crash trap
  for any bad URL)
- app/[section]/page.tsx — redirects to the section's first sub
- app/[section]/[sub]/page.tsx — generateStaticParams over the registry,
  dynamicParams=false (404s unknown slugs), renders REAL_PAGES[section/sub]
  if present else the placeholder SubPageView

Home and changelog stay as static siblings (static wins over the dynamic
[section] segment). Color page components moved app/foundations/_components
→ app/_pages (the foundations dir is gone). Real-page overrides live in
app/_lib/real-pages.tsx.

Verified: all 30 sub-pages + 7 section redirects (307) + home + changelog
prerender; bad URLs 404; rails, breadcrumb, active-nav, theme picker, and
the Color page all intact.
