--- apps/docs/app/agents/[sub]/page.tsx
@@ -1,15 +1,0 @@
-import { SubPageView } from '../../_components/SubPageView';
-import { SECTIONS } from '../../_lib/sections';
-
-export function generateStaticParams() {
-  return SECTIONS.agents.subs.map((s) => ({ sub: s.slug }));
-}
-
-export default async function Page({
-  params,
-}: {
-  params: Promise<{ sub: string }>;
-}) {
-  const { sub } = await params;
-  return <SubPageView sectionSlug="agents" subSlug={sub} />;
-}
