--- apps/docs/app/foundations/page.tsx
@@ -1,8 +1,0 @@
-import { redirect } from 'next/navigation';
-
-import { getDefaultSub } from '../_lib/sections';
-
-export default function Page() {
-  const sub = getDefaultSub('foundations');
-  redirect(`/foundations/${sub}`);
-}
