--- apps/docs/app/getting-started/layout.tsx
@@ -1,7 +1,0 @@
-import { SectionLayout } from '../_components/SectionLayout';
-import { getSection } from '../_lib/sections';
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  const section = getSection('getting-started')!;
-  return <SectionLayout section={section}>{children}</SectionLayout>;
-}
